Clients control different tanks
Rotation consideration needs to be considered.

Either store two rotation values, or keep one but have this be turn based so we can check which tank barrel to rotate either based on values or the current turn. If turn based, we only need to receive and handle one set of mouse loc as we aren't interested in more than one.